Moved Template Node Tree helper class into the Controls folder.
