+++ DataDictionary.Main/Forms/Scripting/Template.Binding.cs
@@ -2,6 +2,7 @@
 using DataDictionary.BusinessLayer.AppSecurity;
 using DataDictionary.BusinessLayer.DbWorkItem;
 using DataDictionary.BusinessLayer.ToolSet;
+using DataDictionary.Main.Controls;
 using DataDictionary.Resource.Enumerations;
 using System.ComponentModel;
 using System.Data;
