+++ DataDictionary.Main/Forms/Scripting/TemplateNode.Binding.cs
@@ -1,6 +1,7 @@
 ﻿using DataDictionary.BusinessLayer.AppScripting;
 using DataDictionary.BusinessLayer.AppSecurity;
 using DataDictionary.BusinessLayer.DbWorkItem;
+using DataDictionary.Main.Controls;
 using System.ComponentModel;
 using System.Data;
 using System.Diagnostics.CodeAnalysis;
